# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenLaw API",
    "description": "REST API for OpenLaw. Also the third-party integration surface (TECH-003).",
    "version": "0.0.1",
    "license": {
      "name": "AGPL-3.0-only",
      "identifier": "AGPL-3.0-only"
    }
  },
  "components": {
    "schemas": {
...
      "Problem": {
        "type": "object",
        "properties": {
          "type": {
            "default": "about:blank",
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": "string"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "path": {
                  "type": "string"
                },
                "message": {
                  "type": "string"
                }
              },
              "required": ["path", "message"],
              "additionalProperties": false
            }
          }
        },
        "required": ["type", "title", "status"],
        "additionalProperties": false
      },
...
    }
  },
  "paths": {
...
    "/api/v1/fields": {
      "get": {
        "operationId": "listFields",
        "summary": "The shared field catalog (CTR-016) scoped to contract and global fields, in creation order; archived rows only with includeArchived=true",
        "tags": ["fields"],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "enum": ["true", "false"]
            },
            "in": "query",
            "name": "includeArchived",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "fields": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "slug": {
                            "type": "string"
                          },
                          "displayName": {
                            "type": "string"
                          },
                          "description": {
                            "anyOf": [
                              {
                                "type": "string"
                              },
                              {
                                "type": "null"
                              }
                            ]
                          },
                          "moduleScope": {
                            "type": "string",
                            "enum": ["contract", "global"]
                          },
                          "fieldType": {
                            "type": "string",
                            "enum": [
                              "text",
                              "long_text",
                              "number",
                              "date",
                              "boolean",
                              "single_select",
                              "multi_select",
                              "user",
                              "entity"
                            ]
                          },
                          "options": {
                            "anyOf": [
                              {
                                "type": "array",
                                "items": {
                                  "type": "string"
                                }
                              },
                              {
                                "type": "null"
                              }
                            ]
                          },
                          "fieldTag": {
                            "type": "string",
                            "enum": ["business", "legal"]
                          },
                          "aiPrompt": {
                            "anyOf": [
                              {
                                "type": "string"
                              },
                              {
                                "type": "null"
                              }
                            ]
                          },
                          "archivedAt": {
                            "anyOf": [
                              {
                                "type": "string",
                                "format": "date-time",
                                "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"
                              },
                              {
                                "type": "null"
                              }
                            ]
                          },
                          "inUseCount": {
                            "type": "integer",
                            "minimum": -9007199254740991,
                            "maximum": 9007199254740991
                          }
                        },
                        "required": [
                          "id",
                          "slug",
                          "displayName",
                          "description",
                          "moduleScope",
                          "fieldType",
                          "options",
                          "fieldTag",
                          "aiPrompt",
                          "archivedAt",
                          "inUseCount"
                        ],
                        "additionalProperties": false
                      }
                    }
                  },
                  "required": ["fields"],
                  "additionalProperties": false
                }
              }
            }
          },
          "default": {
            "description": "Problem details (RFC 9457)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "createField",
        "summary": "Define a field: the slug derives from the name and the field type is picked here, once — both are immutable after creation. Select types take their options list; an AI prompt rides on contract-scoped fields only (CTR-008)",
        "tags": ["fields"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "displayName": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 100
                  },
                  "description": {
                    "type": "string",
                    "maxLength": 500
                  },
                  "moduleScope": {
                    "type": "string",
                    "enum": ["contract", "global"]
                  },
                  "fieldType": {
                    "type": "string",
                    "enum": [
                      "text",
                      "long_text",
                      "number",
                      "date",
                      "boolean",
                      "single_select",
                      "multi_select",
                      "user",
                      "entity"
                    ]
                  },
                  "fieldTag": {
                    "type": "string",
                    "enum": ["business", "legal"]
                  },
                  "options": {
                    "minItems": 1,
                    "maxItems": 100,
                    "type": "array",
                    "items": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 100
                    }
                  },
                  "aiPrompt": {
                    "type": "string",
                    "maxLength": 2000
                  }
                },
                "required": ["displayName", "moduleScope", "fieldType", "fieldTag"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "field": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "slug": {
                          "type": "string"
                        },
                        "displayName": {
                          "type": "string"
                        },
                        "description": {
                          "anyOf": [
                            {
                              "type": "string"
                            },
                            {
                              "type": "null"
                            }
                          ]
                        },
                        "moduleScope": {
                          "type": "string",
                          "enum": ["contract", "global"]
                        },
                        "fieldType": {
                          "type": "string",
                          "enum": [
                            "text",
                            "long_text",
                            "number",
                            "date",
                            "boolean",
                            "single_select",
                            "multi_select",
                            "user",
                            "entity"
                          ]
                        },
                        "options": {
                          "anyOf": [
                            {
                              "type": "array",
                              "items": {
                                "type": "string"
                              }
                            },
                            {
                              "type": "null"
                            }
                          ]
                        },
                        "fieldTag": {
                          "type": "string",
                          "enum": ["business", "legal"]
                        },
                        "aiPrompt": {
                          "anyOf": [
                            {
                              "type": "string"
                            },
                            {
                              "type": "null"
                            }
                          ]
                        },
                        "archivedAt": {
                          "anyOf": [
                            {
                              "type": "string",
                              "format": "date-time",
                              "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"
                            },
                            {
                              "type": "null"
                            }
                          ]
                        },
                        "inUseCount": {
                          "type": "integer",
                          "minimum": -9007199254740991,
                          "maximum": 9007199254740991
                        }
                      },
                      "required": [
                        "id",
                        "slug",
                        "displayName",
                        "description",
                        "moduleScope",
                        "fieldType",
                        "options",
                        "fieldTag",
                        "aiPrompt",
                        "archivedAt",
                        "inUseCount"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": ["field"],
                  "additionalProperties": false
                }
              }
            }
          },
          "default": {
            "description": "Problem details (RFC 9457)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "/"
    }
  ]
}